# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Cashflow Manager API
  description: |
    API for managing personal cashflow transactions, tasks, and activity logs.
    Powers both the web application and Emily (OpenClaw AI agent).
    
    ## Versioning
    - **v1** (`/api/v1/`) - Current version with standardized response envelope
    - **Legacy** (`/api/`) - Deprecated, will be removed in future version
    
    ## Response Format (v1)
    All v1 endpoints return a standardized envelope:
    ```json
    { "success": true, "data": { ... }, "message": "optional" }
    { "success": false, "error": { "code": "...", "message": "..." } }
    ```
  version: 1.1.0
  contact:
    name: Kevin
    email: support@example.com
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://cashflow.example.com
    description: Production server
...
security:
  - PasswordAuth: []
components:
  securitySchemes:
    PasswordAuth:
      type: apiKey
      in: header
      name: X-Password
      description: Shared password for API access
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (15 minute expiry)
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for M2M authentication (prefixed with cfm_)
...
paths:
...
  /api/v1/api-key/generate:
    post:
      tags: [v1 API Key]
      summary: Generate new API key
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, user_id]
              properties:
                name:
                  type: string
                user_id:
                  type: integer
                permissions:
                  type: array
                  items:
                    type: string
                expires_in_days:
                  type: integer
      responses:
        '201':
          description: API key generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  apiKey:
                    type: string
                  name:
                    type: string
                  permissions:
                    type: array
                    items:
                      type: string
                  expires_at:
                    type: string
                    format: date-time
                  message:
                    type: string
  /api/v1/api-key/list:
    get:
      tags: [v1 API Key]
      summary: List API keys for a user
      security: []
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  keys:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        permissions:
                          type: array
                          items:
                            type: string
                        user_id:
                          type: integer
                        expires_at:
                          type: string
                          format: date-time
                        created_at:
                          type: string
                          format: date-time
                        last_used_at:
                          type: string
                          format: date-time
  /api/v1/api-key/revoke/{id}:
    delete:
      tags: [v1 API Key]
      summary: Revoke an API key
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: API key revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string